Created challenges + functioning

+++ database/learn2program.sql
@@ -146,7 +146,7 @@ CREATE TABLE answers (
     success BOOL NOT NULL,
     uId INT NOT NULL,
     eId INT NOT NULL,
-    time INT,
+    time DECIMAL,
     PRIMARY KEY (id),
     FOREIGN KEY (uId) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (eId) REFERENCES exercises(id) ON DELETE CASCADE
